+++ src/identifier.rs
@@ -1,13 +1,13 @@
 use std::error::Error;
 use std::fmt::{Display, Formatter, Result as FmtResult};
 
-/// Convenient type alias for a [`Result`] holding either an [`ArxivId`] or [`ArxivIdError`]
-pub type ArxivIdResult<'a> = Result<ArxivId<'a>, ArxivIdError>;
+/// Convenient type alias for a [`Result`] holding either an [`ArticleId`] or [`ArticleIdError`]
+pub type ArticleIdResult<'a> = Result<ArticleId<'a>, ArticleIdError>;
 
 /// An error that can occur when parsing and validating arXiv identifiers
 #[non_exhaustive]
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ArxivIdError {
+pub enum ArticleIdError {
 	/// Expected the identifier to start with the string literal "arXiv"
 	ExpectedBeginningLiteral,
 	/// Expected to find a `numbervV` component
@@ -20,9 +20,9 @@ pub enum ArxivIdError {
 	InvalidId,
 }
 
-impl Error for ArxivIdError {}
+impl Error for ArticleIdError {}
 
-impl Display for ArxivIdError {
+impl Display for ArticleIdError {
 	fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
 		match self {
 			Self::ExpectedBeginningLiteral => {
@@ -44,22 +44,22 @@ impl Display for ArxivIdError {
 ///
 /// # Examples
 /// ```
-/// use arxiv::ArxivId;
+/// use arxiv::ArticleId;
 ///
-/// let id = ArxivId::try_from("arXiv:2001.00001");
+/// let id = ArticleId::try_from("arXiv:2001.00001");
 /// assert!(id.is_ok());
 /// ```
 ///
 /// [arxiv-docs]: https://info.arxiv.org/help/arxiv_identifier.html
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct ArxivId<'a> {
+pub struct ArticleId<'a> {
 	year: i16,
 	month: i8,
 	number: &'a str,
 	version: ArticleVersion,
 }
 
-impl<'a> ArxivId<'a> {
+impl<'a> ArticleId<'a> {
 	pub const MIN_YEAR: i16 = 2007i16;
 	pub const MAX_YEAR: i16 = 2099i16;
 	pub const MIN_NUM_DIGITS: usize = 4usize;
@@ -70,7 +70,7 @@ impl<'a> ArxivId<'a> {
 	pub(crate) const TOKEN_DOT: char = '.';
 	pub(crate) const TOKEN_VERSION: char = 'v';
 
-	/// This allows manually creating an [`ArxivId`] from the given components without any
+	/// This allows manually creating an [`ArticleId`] from the given components without any
 	/// validation. Only do this if you have already verified that the components are valid.
 	///
 	/// # Safety
@@ -80,9 +80,9 @@ impl<'a> ArxivId<'a> {
 	///
 	/// # Examples
 	/// ```
-	/// use arxiv::{ArticleVersion, ArxivId};
+	/// use arxiv::{ArticleId, ArticleVersion};
 	///
-	/// let id = ArxivId::new(2011, 1, "00001", ArticleVersion::Num(1));
+	/// let id = ArticleId::new(2011, 1, "00001", ArticleVersion::Num(1));
 	/// ```
 	#[inline]
 	pub const fn new(year: i16, month: i8, number: &'a str, version: ArticleVersion) -> Self {
@@ -94,7 +94,7 @@ impl<'a> ArxivId<'a> {
 		}
 	}
 
-	/// This allows manually creating an [`ArxivId`] from the given components without any version
+	/// This allows manually creating an [`ArticleId`] from the given components without any version
 	/// (assuming it is the latest version). Only do this if you have already verified that the
 	/// components are valid:
 	///
@@ -104,16 +104,16 @@ impl<'a> ArxivId<'a> {
 	///
 	/// # Examples
 	/// ```
-	/// use arxiv::{ArticleVersion, ArxivId};
+	/// use arxiv::{ArticleId, ArticleVersion};
 	///
-	/// let id = ArxivId::new_latest(2011, 1, "00001");
+	/// let id = ArticleId::new_latest(2011, 1, "00001");
 	/// ```
 	#[inline]
 	pub const fn new_latest(year: i16, month: i8, id: &'a str) -> Self {
 		Self::new(year, month, id, ArticleVersion::Latest)
 	}
 
-	/// This allows manually creating an [`ArxivId`] from the given components with a specific version.
+	/// This allows manually creating an [`ArticleId`] from the given components with a specific version.
 	/// Only do this if you have already verified that the components are valid:
 	///
 	///  - The year is between the inclusive range of [2007, 2009].
@@ -122,24 +122,24 @@ impl<'a> ArxivId<'a> {
 	///
 	/// # Examples
 	/// ```
-	/// use arxiv::{ArticleVersion, ArxivId};
+	/// use arxiv::{ArticleId, ArticleVersion};
 	///
-	/// let id = ArxivId::new_versioned(2011, 1, "00001", 1);
+	/// let id = ArticleId::new_versioned(2011, 1, "00001", 1);
 	/// assert_eq!(id.version(), ArticleVersion::Num(1));
 	/// ```
 	pub const fn new_versioned(year: i16, month: i8, id: &'a str, version: u8) -> Self {
 		Self::new(year, month, id, ArticleVersion::Num(version))
 	}
 
-	/// This allows manually creating an [`ArxivId`] from the given components with a version, and
+	/// This allows manually creating an [`ArticleId`] from the given components with a version, and
 	/// will also validate each component for correctness. If any component is invalid, it will return
-	/// an [`ArxivIdError`].
+	/// an [`ArticleIdError`].
 	///
 	/// # Examples
 	/// ```
-	/// use arxiv::{ArticleVersion, ArxivId, ArxivIdError};
+	/// use arxiv::{ArticleId, ArticleIdError, ArticleVersion};
 	///
-	/// let id = ArxivId::try_new(2011, 1, "00001", ArticleVersion::Num(1));
+	/// let id = ArticleId::try_new(2011, 1, "00001", ArticleVersion::Num(1));
 	/// assert!(id.is_ok());
 	/// ```
 	pub fn try_new(
@@ -147,37 +147,37 @@ impl<'a> ArxivId<'a> {
 		month: i8,
 		number: &'a str,
 		version: ArticleVersion,
-	) -> ArxivIdResult {
+	) -> ArticleIdResult {
 		if !(Self::MIN_YEAR..=Self::MAX_YEAR).contains(&year) {
-			return Err(ArxivIdError::InvalidYear);
+			return Err(ArticleIdError::InvalidYear);
 		}
 
 		if !(Self::MIN_MONTH..=Self::MAX_MONTH).contains(&month) {
-			return Err(ArxivIdError::InvalidMonth);
+			return Err(ArticleIdError::InvalidMonth);
 		}
 
 		let length_check = (Self::MIN_NUM_DIGITS..=Self::MAX_NUM_DIGITS).contains(&number.len());
 		let digit_check = number.chars().all(|c| c.is_ascii_digit());
 		if !length_check || !digit_check {
-			return Err(ArxivIdError::InvalidId);
+			return Err(ArticleIdError::InvalidId);
 		}
 
 		Ok(Self::new(year, month, number, version))
 	}
 
-	/// This allows manually creating an [`ArxivId`] from the given components without a version
+	/// This allows manually creating an [`ArticleId`] from the given components without a version
 	/// (assuming it is the latest version), and will also validate each component for correctness.
-	/// If any component is invalid, it will return an [`ArxivIdError`].
+	/// If any component is invalid, it will return an [`ArticleIdError`].
 	///
 	/// # Examples
 	/// ```
-	/// use arxiv::{ArxivId, ArxivIdError};
+	/// use arxiv::{ArticleId, ArticleIdError};
 	///
-	/// let id = ArxivId::try_latest(2011, 1, "00001");
+	/// let id = ArticleId::try_latest(2011, 1, "00001");
 	/// assert!(id.is_ok());
 	/// ```
 	#[inline]
-	pub fn try_latest(year: i16, month: i8, number: &'a str) -> ArxivIdResult {
+	pub fn try_latest(year: i16, month: i8, number: &'a str) -> ArticleIdResult {
 		Self::try_new(year, month, number, ArticleVersion::Latest)
 	}
 
@@ -191,9 +191,9 @@ impl<'a> ArxivId<'a> {
 	///
 	/// # Examples
 	/// ```
-	/// use arxiv::ArxivId;
+	/// use arxiv::ArticleId;
 	///
-	/// let id = ArxivId::try_from("arXiv:2304.11188v1").unwrap();
+	/// let id = ArticleId::try_from("arXiv:2304.11188v1").unwrap();
 	/// assert_eq!(id.year(), 2023);
 	/// ```
 	#[must_use]
@@ -206,9 +206,9 @@ impl<'a> ArxivId<'a> {
 	///
 	/// # Examples
 	/// ```
-	/// use arxiv::ArxivId;
+	/// use arxiv::ArticleId;
 	///
-	/// let id = ArxivId::try_from("arXiv:2304.11188v1").unwrap();
+	/// let id = ArticleId::try_from("arXiv:2304.11188v1").unwrap();
 	/// assert_eq!(id.month(), 04);
 	/// ```
 	#[must_use]
@@ -221,9 +221,9 @@ impl<'a> ArxivId<'a> {
 	///
 	/// # Examples
 	/// ```
-	/// use arxiv::ArxivId;
+	/// use arxiv::ArticleId;
 	///
-	/// let id = ArxivId::try_from("arXiv:2304.11188v1").unwrap();
+	/// let id = ArticleId::try_from("arXiv:2304.11188v1").unwrap();
 	/// assert_eq!(id.number(), "11188");
 	/// ```
 	#[must_use]
@@ -236,9 +236,9 @@ impl<'a> ArxivId<'a> {
 	///
 	/// # Examples
 	/// ```
-	/// use arxiv::{ArticleVersion, ArxivId};
+	/// use arxiv::{ArticleId, ArticleVersion};
 	///
-	///  let id = ArxivId::try_from("arXiv:2304.11188v1").unwrap();
+	///  let id = ArticleId::try_from("arXiv:2304.11188v1").unwrap();
 	/// assert_eq!(id.version(), ArticleVersion::Num(1));
 	/// ```
 	#[must_use]
@@ -251,9 +251,9 @@ impl<'a> ArxivId<'a> {
 	///
 	/// # Examples
 	/// ```
-	/// use arxiv::{ArticleVersion, ArxivId};
+	/// use arxiv::{ArticleId, ArticleVersion};
 	///
-	/// let mut id = ArxivId::try_from("arXiv:2001.00001").unwrap();
+	/// let mut id = ArticleId::try_from("arXiv:2001.00001").unwrap();
 	/// assert_eq!(id.version(), ArticleVersion::Latest);
 
 	/// id.set_version(1);
@@ -268,9 +268,9 @@ impl<'a> ArxivId<'a> {
 	///
 	/// # Examples
 	/// ```
-	/// use arxiv::{ArticleVersion, ArxivId};
+	/// use arxiv::{ArticleId, ArticleVersion};
 	///
-	/// let mut id = ArxivId::try_from("arXiv:2001.00001v1").unwrap();
+	/// let mut id = ArticleId::try_from("arXiv:2001.00001v1").unwrap();
 	/// assert_eq!(id.version(), ArticleVersion::Num(1));
 	///
 	/// id.set_latest();
@@ -282,7 +282,7 @@ impl<'a> ArxivId<'a> {
 	}
 }
 
-impl<'a> Display for ArxivId<'a> {
+impl<'a> Display for ArticleId<'a> {
 	fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
 		let mut year_str = self.year.to_string();
 		let (_, half_year) = year_str.as_mut_str().split_at(2);
@@ -295,19 +295,19 @@ impl<'a> Display for ArxivId<'a> {
 	}
 }
 
-impl<'a> TryFrom<&'a str> for ArxivId<'a> {
-	type Error = ArxivIdError;
+impl<'a> TryFrom<&'a str> for ArticleId<'a> {
+	type Error = ArticleIdError;
 
 	fn try_from(value: &'a str) -> Result<Self, Self::Error> {
-		use ArxivIdError::*;
+		use ArticleIdError::*;
 
 		// break down the arxiv string into its components
-		let parts: Vec<&str> = value.split(ArxivId::TOKEN_COLON).collect();
+		let parts: Vec<&str> = value.split(ArticleId::TOKEN_COLON).collect();
 		if parts.len() != 2 || parts[0] != "arXiv" {
 			return Err(ExpectedBeginningLiteral);
 		}
 
-		let inner_parts: Vec<&str> = parts[1].split(ArxivId::TOKEN_DOT).collect();
+		let inner_parts: Vec<&str> = parts[1].split(ArticleId::TOKEN_DOT).collect();
 		if inner_parts.len() != 2 {
 			return Err(ExpectedNumberVv);
 		}
@@ -319,7 +319,7 @@ impl<'a> TryFrom<&'a str> for ArxivId<'a> {
 		let year = date[0..2].parse::<i16>().map_err(|_| InvalidYear)?;
 		let month = date[2..4].parse::<i8>().map_err(|_| InvalidMonth)?;
 		let (number, version) = parse_numbervv(numbervv).ok_or(ExpectedNumberVv)?;
-		ArxivId::try_new(year + 2000i16, month, number, version)
+		ArticleId::try_new(year + 2000i16, month, number, version)
 	}
 }
 
@@ -371,7 +371,7 @@ pub(crate) fn parse_numbervv(s: &str) -> Option<(&str, ArticleVersion)> {
 	if s.len() > number.len() {
 		let after_number = &mut s[number.len()..].chars().peekable();
 		if after_number
-			.next_if(|c| *c == ArxivId::TOKEN_VERSION)
+			.next_if(|c| *c == ArticleId::TOKEN_VERSION)
 			.is_some()
 		{
 			let consume = after_number
@@ -387,17 +387,17 @@ pub(crate) fn parse_numbervv(s: &str) -> Option<(&str, ArticleVersion)> {
 
 #[cfg(test)]
 mod test_display {
-	use super::ArxivId;
+	use super::ArticleId;
 
 	#[test]
 	fn with_version() {
-		let id = ArxivId::new_versioned(2007, 1, "0001", 1);
+		let id = ArticleId::new_versioned(2007, 1, "0001", 1);
 		assert_eq!(id.to_string(), "arXiv:0701.0001v1");
 	}
 
 	#[test]
 	fn without_version() {
-		let id = ArxivId::new_latest(2007, 1, "0001");
+		let id = ArticleId::new_latest(2007, 1, "0001");
 		assert_eq!(id.to_string(), "arXiv:0701.0001");
 	}
 }
@@ -420,7 +420,7 @@ mod tests_parse_ok {
 
 	#[test]
 	fn from_readme() {
-		let id = ArxivId::try_from("arXiv:0706.0001v1").unwrap();
+		let id = ArticleId::try_from("arXiv:0706.0001v1").unwrap();
 		assert_eq!(id.year, 2007);
 		assert_eq!(id.month, 6);
 		assert_eq!(id.number(), "0001");
@@ -429,31 +429,31 @@ mod tests_parse_ok {
 
 	#[test]
 	fn without_version() {
-		let id = ArxivId::try_from("arXiv:1501.00001");
-		assert_eq!(id, Ok(ArxivId::new_latest(2015, 1, "00001")));
+		let id = ArticleId::try_from("arXiv:1501.00001");
+		assert_eq!(id, Ok(ArticleId::new_latest(2015, 1, "00001")));
 	}
 
 	#[test]
 	fn with_version() {
-		let id = ArxivId::try_from("arXiv:9912.12345v2");
-		assert_eq!(id, Ok(ArxivId::new(2099, 12, "12345", ArticleVersion::Num(2))))
+		let id = ArticleId::try_from("arXiv:9912.12345v2");
+		assert_eq!(id, Ok(ArticleId::new(2099, 12, "12345", ArticleVersion::Num(2))))
 	}
 
 	#[test]
 	fn with_number_4_digits() {
-		let id1 = ArxivId::new_latest(2014, 1, "7878");
+		let id1 = ArticleId::new_latest(2014, 1, "7878");
 		assert_eq!(id1.to_string(), String::from("arXiv:1401.7878"));
 
-		let id2 = ArxivId::new_latest(2014, 12, "7878");
+		let id2 = ArticleId::new_latest(2014, 12, "7878");
 		assert_eq!(id2.to_string(), String::from("arXiv:1412.7878"));
 	}
 
 	#[test]
 	fn with_number_5_digits() {
-		let id1 = ArxivId::new_latest(2014, 1, "00008");
+		let id1 = ArticleId::new_latest(2014, 1, "00008");
 		assert_eq!(id1.to_string(), String::from("arXiv:1401.00008"));
 
-		let id2 = ArxivId::new_latest(2014, 12, "00008");
+		let id2 = ArticleId::new_latest(2014, 12, "00008");
 		assert_eq!(id2.to_string(), String::from("arXiv:1412.00008"));
 	}
 }
@@ -464,31 +464,31 @@ mod tests_parse_err {
 
 	#[test]
 	fn empty_string() {
-		let id = ArxivId::try_from("");
-		assert_eq!(id, Err(ArxivIdError::ExpectedBeginningLiteral));
+		let id = ArticleId::try_from("");
+		assert_eq!(id, Err(ArticleIdError::ExpectedBeginningLiteral));
 	}
 
 	#[test]
 	fn no_numbervv() {
-		let id = ArxivId::try_from("arXiv:1501");
-		assert_eq!(id, Err(ArxivIdError::ExpectedNumberVv));
+		let id = ArticleId::try_from("arXiv:1501");
+		assert_eq!(id, Err(ArticleIdError::ExpectedNumberVv));
 	}
 
 	#[test]
 	fn invalid_year() {
-		let maybe_id = ArxivId::try_latest(2006, 1, "00001");
-		assert_eq!(maybe_id, Err(ArxivIdError::InvalidYear));
+		let maybe_id = ArticleId::try_latest(2006, 1, "00001");
+		assert_eq!(maybe_id, Err(ArticleIdError::InvalidYear));
 	}
 
 	#[test]
 	fn invalid_month() {
-		let maybe_id = ArxivId::try_latest(2007, i8::MAX, "00001");
-		assert_eq!(maybe_id, Err(ArxivIdError::InvalidMonth));
+		let maybe_id = ArticleId::try_latest(2007, i8::MAX, "00001");
+		assert_eq!(maybe_id, Err(ArticleIdError::InvalidMonth));
 	}
 
 	#[test]
 	fn invalid_id() {
-		let maybe_id = ArxivId::try_latest(2007, 11, "");
-		assert_eq!(maybe_id, Err(ArxivIdError::InvalidId));
+		let maybe_id = ArticleId::try_latest(2007, 11, "");
+		assert_eq!(maybe_id, Err(ArticleIdError::InvalidId));
 	}
 }